Update fixes-dates.php
Modified line 171 This change adds a check to ensure $post is an object before accessing its ID property. If $post is not an object, $post_id will be set to null, avoiding the warning.

Previously $post->ID was being accessed, but it appears that $post might be an integer or something other than an object, leading to the warning.